Fix styles
